requirements.txt
# requirements.txt
requests
pyserial
